python/requirements.txt
pytesseract
Pillow
groq
easyocr
python-dotenv
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
